reverted to 053e7585bd58cf4cce3cfdf8b964bd80e5daf037

+++ bottomLayer/src/routes/private/user.ts
@@ -1,12 +1,8 @@
 import express, { type Request, type Response } from 'express';
 import { pool } from '../../utils/sqlImport';
-import {
-  responseCallbackDelete,
-  responseCallbackPost,
-  responseCallbackUpdate,
-  responseCallbackGet
-} from '../../utils/responseCallback';
+import { responseCallbackDelete, responseCallbackPost, responseCallbackUpdate, responseCallbackGet } from '../../utils/responseCallback';
 import { checkAuthenticated } from '../../middleware/auth';
+import { convertImage } from '../../s3/convertImage';
 const router = express.Router();
 
 router.get('/', (req: Request, res: Response): void => {
@@ -14,10 +10,7 @@ router.get('/', (req: Request, res: Response): void => {
 
   const getUser = async (): Promise<void> => {
     try {
-      const user = await pool.query(
-        'SELECT first_name, last_name, email, username, profile_picture FROM backend_schema.user WHERE uid = $1',
-        [userId]
-      );
+      const user = await pool.query('SELECT uid, first_name, last_name, email, username, profile_picture FROM backend_schema.user WHERE uid = $1', [userId]);
       const result = user.rows[0];
       responseCallbackGet(null, result, res, 'User');
     } catch (error) {
@@ -43,25 +36,13 @@ router.post('/', checkAuthenticated, (req: Request, res: Response) => {
 
   const insertUser = async (): Promise<void> => {
     try {
-      // const URL = await convertImage(profile_picture, username);
-      await pool.query(
-        `
+      const URL = await convertImage(profile_picture, username);
+      await pool.query(`
       INSERT INTO backend_schema.user (
         first_name, last_name, email, username, password, private_option, followers, following, profile_picture
         ) VALUES ( 
           $1, $2, $3, $4, $5, $6, $7, $8, $9)`,
-        [
-          first_name,
-          last_name,
-          email,
-          username,
-          password,
-          private_option,
-          followers,
-          following,
-          profile_picture
-        ]
-      );
+      [first_name, last_name, email, username, password, private_option, followers, following, URL]);
       responseCallbackPost(null, res, 'User');
     } catch (error) {
       responseCallbackPost(error, res);
@@ -79,10 +60,7 @@ router.delete('/', checkAuthenticated, (req: Request, res: Response): void => {
 
   const deleteUser = async (): Promise<void> => {
     try {
-      const deleteUser = await pool.query(
-        'DELETE FROM backend_schema.user WHERE uid = $1',
-        [userId]
-      );
+      const deleteUser = await pool.query('DELETE FROM backend_schema.user WHERE uid = $1', [userId]);
       responseCallbackDelete(null, userId, res, 'User', deleteUser.rowCount);
     } catch (error) {
       responseCallbackDelete(error, userId, res, 'User');
@@ -111,9 +89,8 @@ router.put('/', checkAuthenticated, (req: Request, res: Response): void => {
   } = req.body;
   const updateUser = async (): Promise<void> => {
     try {
-      // const URL = await convertImage(profile_picture, username);
-      const updateUser = await pool.query(
-        `UPDATE backend_schema.user
+      const URL = await convertImage(profile_picture, username);
+      const updateUser = await pool.query(`UPDATE backend_schema.user
         SET first_name = $1,
             last_name = $2,
             email = $3,
@@ -124,19 +101,7 @@ router.put('/', checkAuthenticated, (req: Request, res: Response): void => {
             following = $8,
             profile_picture = $9
         WHERE uid = $10`,
-        [
-          first_name,
-          last_name,
-          email,
-          username,
-          password,
-          private_option,
-          followers,
-          following,
-          profile_picture,
-          userId
-        ]
-      );
+      [first_name, last_name, email, username, password, private_option, followers, following, URL, userId]);
       responseCallbackUpdate(null, userId, res, 'User', updateUser.rowCount);
     } catch (error) {
       responseCallbackUpdate(error, userId, res, 'User');
